==> RMDs/Microglia/26_Merged_FGSEA.Rmd <==
---
title: "26_Merged_FGSEA"
author: "Nathan Haywood"
date: "2025-06-17"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
# Clean environment
rm(list = ls(all.names = TRUE)) # will clear all objects including hidden objects
gc() # free up memory and report the memory usage
options(max.print = .Machine$integer.max, scipen = 999, stringsAsFactors = F, dplyr.summarise.inform = F) # avoid truncated output in R console and scientific notation

# Set seed
set.seed(123456)

# Loading relevant libraries 
library(tidyverse) # includes ggplot2, for data visualisation. dplyr, for data manipulation.
library(RColorBrewer) # for a colourful plot
library(fgsea)
```


```{r}
# Functions ===================================================
## Function: Adjacency matrix to list -------------------------
matrix_to_list <- function(pws){
  pws.l <- list()
  for (pw in colnames(pws)) {
    pws.l[[pw]] <- rownames(pws)[as.logical(pws[, pw])]
  }
  return(pws.l)
}

## Function: prepare_gmt --------------------------------------
prepare_gmt <- function(gmt_file, genes_in_data, savefile = FALSE){
  # for debug
  #file <- gmt_files[1]
  #genes_in_data <- df$gene_symbol
  
  # Read in gmt file
  gmt <- gmtPathways(gmt_file)
  hidden <- unique(unlist(gmt))
  
  # Convert gmt file to a matrix with the genes as rows and for each go annotation (columns) the values are 0 or 1
  mat <- matrix(NA, dimnames = list(hidden, names(gmt)),
                nrow = length(hidden), ncol = length(gmt))
  for (i in 1:dim(mat)[2]){
    mat[,i] <- as.numeric(hidden %in% gmt[[i]])
  }
  
  #Subset to the genes that are present in our data to avoid bias
  hidden1 <- intersect(genes_in_data, hidden)
  mat <- mat[hidden1, colnames(mat)[which(colSums(mat[hidden1,])>5)]] # filter for gene sets with more than 5 genes annotated
  # And get the list again
  final_list <- matrix_to_list(mat) # for this we use the function we previously defined
  
  if(savefile){
    saveRDS(final_list, file = paste0(gsub('.gmt', '', gmt_file), '_subset_', format(Sys.time(), '%d%m'), '.RData'))
  }
  
  print('Yahoo! .gmt conversion successfull!:)')
  return(final_list)
}

# Analysis ====================================================

## 1. Read in data -----------------------------------------------------------
setwd('/Users/nate/Desktop/Mikes_ATF3_Data/CSVs/')
df <- read.csv("Merged_DEGs_by_Cluster_by_Genotype.csv", row.names = 1)
df <- subset(df, df$clusterID == "1")

## 2. Prepare background genes -----------------------------------------------

# Download gene sets .gmt files
#https://www.gsea-msigdb.org/gsea/msigdb/collections.jsp

# For GSEA
# Filter out the gmt files for KEGG, Reactome and GOBP
my_genes <- df$gene

setwd('/Users/nate/Desktop/Mikes_ATF3_Data/MSIGDB_GMTs/')
gmt_files <- list.files(path = getwd(), pattern = '.gmt', full.names = TRUE)
gmt_files
bg_genes <- prepare_gmt(gmt_files[1], df$gene, savefile = FALSE)

# 1: Canonical pathways, 2: Predicted TFs, 3: Immunology 4: Hallmark genes,
```

```{r}
rankings <- sign(df$logFC)*(-log10(df$PValue)) # we will use the signed p values from spatial DGE as ranking
names(rankings) <- df$gene # genes as names#


head(rankings)

max_ranking <- max(rankings[is.finite(rankings)])
min_ranking <- min(rankings[is.finite(rankings)])
rankings <- replace(rankings, rankings > max_ranking, max_ranking * 10)
rankings <- replace(rankings, rankings < min_ranking, min_ranking * 10)

rankings <- sort(rankings, decreasing = TRUE) # sort genes by ranking
plot(rankings)

max(rankings)
min(rankings)
```


```{r}
ggplot(data.frame(gene_symbol = names(rankings)[1:50], ranks = rankings[1:50]), aes(gene_symbol, ranks)) + 
	geom_point() +
	theme_classic() + 
	theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
```


```{r}
## 4. Run GSEA ---------------------------------------------------------------
# Easy peasy! Run fgsea with the pathways 
GSEAres <- fgsea(pathways = bg_genes, # List of gene sets to check
                 stats = rankings,
                 scoreType = 'std', # in this case we have both pos and neg rankings. if only pos or neg, set to 'pos', 'neg'
                 minSize = 10,
                 maxSize = 500,
                 nproc = 1) # for parallelisation


head(GSEAres)
sum(GSEAres[, padj < 0.01])
sum(GSEAres[, pval < 0.01])

topPathwaysUp <- GSEAres[ES > 0][head(order(padj), n = 25), pathway]
topPathwaysDown <- GSEAres[ES < 0][head(order(padj), n = 25), pathway]
topPathways <- c(topPathwaysUp, rev(topPathwaysDown))
#pdf(file = paste0(filename, '_gsea_top30pathways.pdf'), width = 20, height = 15)
plotGseaTable(bg_genes[topPathways], stats = rankings, fgseaRes = GSEAres, gseaParam = 0.5)
#dev.off()

# Select only independent pathways, removing redundancies/similar pathways
collapsedPathways <- collapsePathways(GSEAres[order(padj)][padj < 0.05], bg_genes, rankings)
mainPathways <- GSEAres[pathway %in% collapsedPathways$mainPathways][order(-NES), pathway]
#pdf(file = paste0('GSEA/Selected_pathways/', paste0(filename, background_genes, '_gsea_mainpathways.pdf')), width = 20, height = 15)
plotGseaTable(bg_genes[mainPathways], rankings, GSEAres, gseaParam = 0.5)
#dev.off()

# plot the most significantly enriched pathway
plotEnrichment(bg_genes[[head(GSEAres[order(padj), ], 1)$pathway]],
               rankings) + 
  labs(title = head(GSEAres[order(padj), ], 1)$pathway)

## 5. Save the results -----------------------------------------------
name_of_comparison <- 'ATF3KO_DEGs'
#background_genes <- 'gobp'
#filename <- paste0(out_path, 'GSEA/', name_of_comparison, '_', background_genes) 
#saveRDS(GSEAres, file = paste0(filename, '_gsea_results.RDS'))
#data.table::fwrite(GSEAres, file = paste0(filename, '_gsea_results.tsv'), sep = "\t", sep2 = c("", " ", ""))
```


```{r}
library(ComplexHeatmap)
library(fgsea)
library(babelgene)
library(dplyr)
library(data.table)

extract_gene<-function(df) {
  output<-df$gene
  output
} #simple function to extract gene column as vector
extract_Gene<-function(df) {
  output<-df$Gene
  output
} #simple function to extract gene column as vector


setwd('/Users/nate/Desktop/Mikes_ATF3_Data/CSVs/')
macs.markers <- read.csv("Merged_DEGs_by_Cluster_by_Genotype.csv")

#STEP5: GSEA analysis of microglial gene signatures ####
mast.test.filtered<-dplyr::select(macs.markers, gene, logFC, clusterID)
mast.test.filtered<-arrange(mast.test.filtered, clusterID, logFC) #This ordering is critical to get correct results 

#Pull in Vahid's DEG genes to create pathway for each inferred MG cell-type 
#Vahid.DEGs<-qread("SubstateMG_w1_human_iNPH_.qs") #This is the states broken down in more detail 

setwd('/Users/nate/Desktop/Mikes_ATF3_Data/MSIGDB_GMTs/')
gmt_files <- list.files(path = getwd(), pattern = '.gmt', full.names = TRUE)
gmt_files
bg_genes <- prepare_gmt(gmt_files[1], df$gene, savefile = FALSE)
Mike.DEGs <- bg_genes[mainPathways]


Mike.DEGs<-filter(Mike.DEGs, FDR<0.05)
Mike.DEGs<-filter(Mike.DEGs, logFC>0) #Just positively expressed genes
#Mike.DEGs<-filter(Mike.DEGs, cluster_name!="Myeloid_CD300E", cluster_name!="Macrophage_CD200R1")
Mike.DEGs<-arrange(Mike.DEGs, comparison, desc(logFC)) #This ordering is critical to get correct results 
#Vahid.DEGs<-filter(Vahid.DEGs, avg_logFC>0.1)

Mike.DEGs<-split(Mike.DEGs,Mike.DEGs$comparison)
Mike.DEGs.list<-lapply(Mike.DEGs, extract_gene)

#Use Babelgene to convert mouse to human 
lapply(Mike.DEGs.list, length)
#elements_to_remove <- c("Myeloid_CD300E", "Macrophage_CD200R1")
#Vahid.DEGs.list <- Vahid.DEGs.list[setdiff(names(Vahid.DEGs.list), elements_to_remove)]
#Mike.DEGs.list <- lapply(Mike.DEGs.list, function(x) orthologs(genes = x, species = "mouse", human = TRUE, top=TRUE))
#Mike.DEGs.list <- lapply(Mike.DEGs.list, function(x) x$symbol)
#lapply(Mike.DEGs.list, length)

#Mike>DEGs.list <- bg_genes

#Run GSEA on positively expressed, signifncant DEGs from all clusters in Gerrits using fgsea (turn this into a function?)
num_clusters<-length(unique(mast.test.filtered$cluster))
fgseaRes<-data.table()

for(i in 1:num_clusters) { 
 cluster.test<-unique(mast.test.filtered$clusterID)[i] #Clusters are 0 indexed in Seurat and Liger 
 cluster.genes<-dplyr::filter(mast.test.filtered, clusterID==cluster.test)
 cluster.genes.vector<-cluster.genes$logFC
 names(cluster.genes.vector)<-cluster.genes$gene
 fgseaRes.temp <- fgsea(pathways = Mike.DEGs.list ,nperm = 10000
             ,stats = cluster.genes.vector, minSize = 1)
 fgseaRes.temp$clusterID<-cluster.test
 fgseaRes<-rbind(fgseaRes, fgseaRes.temp)
}
fgseaRes 

fgseaRes<- dplyr::mutate(fgseaRes, p_sign = sign(NES) * -log10(padj))
gsea_cast <- dcast(fgseaRes, clusterID ~ pathway, value.var = "p_sign")

p_star <- matrix(ncol= ncol(gsea_cast[,-1]), nrow = nrow(gsea_cast))
p_star[gsea_cast[,-1] > -log10(0.05) | gsea_cast[,-1] < log10(0.05)] <- "*"
p_star[gsea_cast[,-1] > -log10(0.01) | gsea_cast[,-1] < log10(0.01)] <- "**"
p_star[gsea_cast[,-1] > -log10(0.001) | gsea_cast[,-1] < log10(0.001)] <- "***"
p_star[gsea_cast[,-1] > -log10(0.0001) | gsea_cast[,-1] < log10(0.0001)] <- "****"
p_star[is.na(p_star)] <- ""

mat<-as.matrix(gsea_cast[,-1]) #Create a matrix and add clusters as rownames 
rownames(mat)<-gsea_cast$clusterID

Heatmap(mat, cluster_rows = T, cluster_columns = T, border = TRUE
    , row_title_side = "right", column_title = ""
    , column_title_side = "bottom", cell_fun = function(j, i, x, y, width, height, fill) {
     grid.text(sprintf("%s", p_star[i, j]), x, y, gp = gpar(fontsize = 10))
    }, name = "ScaledExp")

```
